requirements.txt
aioboto3==11.3.0
asyncpg==0.28.0
boto3==1.28.17
celery==5.3.4
fastapi==0.103.0
langchain==0.0.288
openai==0.28.0
pgvector==0.2.2
psycopg2==2.9.7
pypdf==3.16.0
python-dotenv==1.0.0
python-jose==3.3.0
requests==2.31.0
SQLAlchemy==2.0.20
starlette==0.27.0
tiktoken==0.5.1
uvicorn==0.23.2
yookassa==2.4.0
python-multipart==0.0.6
jose==1.0.0
